# perf: reduce `bigint` parsing overhead