--- a/app/json.ts
+++ b/app/json.ts
@@ -5,8 +5,7 @@ import {
   useLoaderData as useRemixLoaderData,
   useMatches as useRemixMatches,
 } from 'remix';
-import { deserialize, serialize } from 'superjson';
-import { SuperJSONResult } from 'superjson/dist/types';
+import { parse, stringify } from 'json-bigint';
 
 type JsonResponse = ReturnType<typeof remixJson>;
 type MetaArgs = Parameters<MetaFunction>[0];
@@ -20,30 +19,25 @@ export function json<Data>(
   obj: Data,
   init?: number | ResponseInit
 ): JsonResponse {
-  const superJsonResult = serialize(obj);
-  return remixJson(superJsonResult, init);
-}
-
-export function parse<Data>(superJsonResult: SuperJSONResult) {
-  return deserialize<Data>(superJsonResult);
+  return remixJson(stringify(obj), init);
 }
 
 export function withSuperJSON<Data>(
   metaFn: MetaFunction
 ): SuperJSONMetaFunction<Data> {
   return ({ data, ...rest }: MetaArgs): HtmlMetaDescriptor =>
-    metaFn({ ...rest, data: parse<Data>(data as SuperJSONResult) });
+    metaFn({ ...rest, data: parse(data as string) as Data });
 }
 
 export function useLoaderData<Data>(): Data {
-  const loaderData = useRemixLoaderData<SuperJSONResult>();
-  return parse<Data>(loaderData);
+  const loaderData = useRemixLoaderData<string>();
+  return parse(loaderData) as Data;
 }
 
 export function useMatches(): ReturnType<typeof useRemixMatches> {
   const matches = useRemixMatches();
   return matches.map(({ data, ...rest }) => ({
     ...rest,
-    data: data ? parse(data as SuperJSONResult) : data,
+    data: data ? parse(data as unknown as string) : data,
   }));
 }