Context from [the Remix Discord server](https://discord.com/channels/770287896669978684/770287896669978687/961351028219969617) (invite [here](https://discord.gg/remix)):

> Alright, I found my issue. I had a super nested `<Action>` component that used `useMatches()` which caused its render to take 43ms (v.s. 1ms when that `useMatches()` call was removed); because I had a couple hundred of these `<Action>` components in my React tree, rendering took 7s. I inspected the result of `useMatches()` and it seemed to be the same throughout the render which makes me wonder why it caused such a large increase in render time. Is there something that I'm missing here about the way that `useMatches()` works internally that causes calling it to add so much overhead to an otherwise small component render (i.e. 43ms v.s. 1ms)?

> Actually, nevermind about the `useMatches()` thing either. Turns out I'm wrapping that `useMatches()` call with a bunch of `superjson` stuff which is quite slow.

> I think my performance issue stemmed from the fact that I do have slow code (as I have to wrap Remix's `json`, `useLoaderData()`, and `useMatches` helpers with `superjson` in order to support the `BigInt`s that are coming from my database). I was using `superjson-remix`, but that's quite slow so I'm thinking I'll use `json-bigint` but that requires that I have access to the response data raw JSON string (instead of the parsed result that `useLoaderData()` gives me). Is there a React hook that Remix has that can expose that raw response data string v.s. the parsed JSON object?